+++ src/matrix.cpp
@@ -4,26 +4,26 @@
 #include <iostream>
 #include <random>
 
-LPP::Matrix::Matrix(const std::vector<std::vector<double>>& m) {entries = m;}
+LPP::Matrix::Matrix(const std::vector<std::vector<float>>& m) {entries = m;}
 
-LPP::Matrix::Matrix(std::vector<std::vector<double>>& m) {entries = std::move(m);}
+LPP::Matrix::Matrix(std::vector<std::vector<float>>& m) {entries = std::move(m);}
 
 LPP::Matrix::Matrix(const size_t rows, const size_t cols): entries{rows}
 {
     if (rows < LPP::CONSTANTS::MATRIX_PARALLEL_THRESHOLD) {
         for (size_t r = 0; r < rows; r++) {
-            entries[r] = std::vector<double>(cols, 0.0);
+            entries[r] = std::vector<float>(cols, 0.0);
         }
     } else {
         LPP::matrix_parallel_init(entries, rows, cols, 0.0);
     }
 }
 
-LPP::Matrix::Matrix(const size_t rows, const size_t cols, const double c): entries{rows}
+LPP::Matrix::Matrix(const size_t rows, const size_t cols, const float c): entries{rows}
 {
     if (rows < LPP::CONSTANTS::MATRIX_PARALLEL_THRESHOLD) {
         for (size_t r = 0; r < rows; r++) {
-            entries[r] = std::vector<double>(cols, c);
+            entries[r] = std::vector<float>(cols, c);
         }
     } else {
         LPP::matrix_parallel_init(entries, rows, cols, c);
@@ -60,7 +60,7 @@ void LPP::Matrix::print_entries() const
     std::cout << '}' << std::endl;
 }
 
-double LPP::Matrix::get(size_t i, size_t j) const
+float LPP::Matrix::get(size_t i, size_t j) const
 {
     if (i >= entries.size() || j >= entries[0].size()) {
         throw std::invalid_argument("Vector entry extraction out of range");
@@ -68,7 +68,7 @@ double LPP::Matrix::get(size_t i, size_t j) const
     return entries[i][j];
 }
 
-void LPP::Matrix::set(size_t i, size_t j, double s)
+void LPP::Matrix::set(size_t i, size_t j, float s)
 {
     if (i >= entries.size() || j >= entries[0].size()) {
         throw std::invalid_argument("Vector entry setting out of range");
@@ -76,11 +76,11 @@ void LPP::Matrix::set(size_t i, size_t j, double s)
     entries[i][j] = s;
 }
 
-const std::vector<double>& LPP::Matrix::operator[](size_t i) const {return entries[i];}
+const std::vector<float>& LPP::Matrix::operator[](size_t i) const {return entries[i];}
 
-std::vector<double>& LPP::Matrix::operator[](size_t i) {return entries[i];}
+std::vector<float>& LPP::Matrix::operator[](size_t i) {return entries[i];}
 
-std::vector<double> LPP::Matrix::operator*(const std::vector<double>& v) const
+std::vector<float> LPP::Matrix::operator*(const std::vector<float>& v) const
 {
     if (entries.empty()) {
         const auto msg = "Attempting to multiply by empty matrix";
@@ -91,7 +91,7 @@ std::vector<double> LPP::Matrix::operator*(const std::vector<double>& v) const
         throw std::invalid_argument(msg);
     }
 
-    std::vector<double> res(entries.size());
+    std::vector<float> res(entries.size());
     for (size_t i = 0; i < res.size(); i++) {
         res[i] = entries[i] * v;
     }
@@ -130,14 +130,14 @@ LPP::Matrix& LPP::Matrix::operator-=(const Matrix& m)
     return *this;
 }
 
-void LPP::matrix_mult_helper(Matrix& m1, const size_t start, const size_t end, const double c)
+void LPP::matrix_mult_helper(Matrix& m1, const size_t start, const size_t end, const float c)
 {
     for (size_t r = start; r < end; r++) {
         m1[r] *= c;
     }
 }
 
-LPP::Matrix& LPP::Matrix::operator*=(const double c)
+LPP::Matrix& LPP::Matrix::operator*=(const float c)
 {
     if (this->rows() < LPP::CONSTANTS::MATRIX_PARALLEL_THRESHOLD) {
         for (size_t i = 0; i < entries.size(); i++) {
@@ -163,10 +163,10 @@ bool LPP::same_dims(const Matrix& m1, const Matrix& m2)
     return m1.rows() == m2.rows() && m1.cols() == m2.cols();
 }
 
-void LPP::print_object(const std::vector<double>& v)
+void LPP::print_object(const std::vector<float>& v)
 {
     std::cout << '{';
-    for (const double d : v) {
+    for (const float d : v) {
         std::cout << d << ", ";
     }
     std::cout << '}' << std::endl;
